BI_board/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Celery and Redis
celery==5.3.4
django-celery-results==2.5.1
redis==5.0.1
django-redis==5.4.0

# Data Processing and ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
openpyxl==3.1.2

# Fuzzy matching for column mapping
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Stripe for payments
stripe==7.8.0

# Environment variables
python-dotenv==1.0.0

# Nested routers for DRF
djangorestframework-nested==0.0.1

# Additional utilities
Pillow==10.1.0
requests==2.31.0
